:wrench: maintenance

+++ uninstall.php
@@ -3,5 +3,22 @@
 // ensure this plugin is actually being uninstalled
 defined( 'WP_UNINSTALL_PLUGIN' ) or exit();
 
-delete_option( 'gnetwork_blog' ); // FIXME: it just for main blog!
+if ( is_multisite() ) {
+
+	global $wpdb;
+
+	if ( $blogs = $wpdb->get_results( "SELECT blog_id FROM {$wpdb->blogs}", ARRAY_A ) ) {
+
+	 	foreach ( $blogs as $blog ) {
+			switch_to_blog( $blog['blog_id'] );
+			delete_option( 'gnetwork_blog' );
+		}
+
+		restore_current_blog();
+	}
+
+} else {
+	delete_option( 'gnetwork_blog' );
+}
+
 delete_site_option( 'gnetwork_site' );
